Updates to switch API to Postgres instead of Mongo (/galleries route).
